requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.10.6
python-dotenv~=1.0.1
dotenv~=0.9.9
h11~=0.14.0
distro~=1.9.0
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.8
starlette~=0.46.1
requests~=2.32.3
anyio~=4.8.0
annotated-types~=0.7.0
httpx~=0.28.1
openai~=1.65.4
pytest~=8.3.5
attrs~=25.3.0
future~=0.18.2
httpcore~=1.0.7
idna~=3.10
jiter~=0.8.2
websockets~=14.2
realtime~=2.4.2
certifi~=2025.1.31
h2~=4.2.0
supabase~=2.15.0
hpack~=4.1.0
hyperframe~=6.1.0
pip~=21.2.4
wheel~=0.37.0
multidict~=6.3.2
propcache~=0.3.1
PyJWT~=2.10.1
pluggy~=1.5.0
iniconfig~=2.1.0
yarl~=1.19.0
async-timeout~=5.0.1
aiohttp~=3.11.16
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.6.1
six~=1.15.0
python-dateutil~=2.9.0.post0
storage3~=0.11.3
gotrue~=2.12.0
postgrest~=1.0.1
supafunc~=0.9.4
StrEnum~=0.4.15
packaging~=24.2
deprecation~=2.1.0
pytest-mock~=3.14.0
anthropic~=0.49.0
protobuf~=6.30.2
urllib3~=2.3.0
tqdm~=4.67.1
charset-normalizer~=3.4.1
google-genai~=1.10.0